# api/openapi.yaml
openapi: 3.1.2
info:
  title: Synodus Core API
  version: 1.0.0
  description: >-
    The reviewed v1 browser API contract. Organization, team, and task
    operations remain dark until the D06 route cutover.
...
paths:
...
  /api/v1/auth/session:
    get:
      tags: [Authentication]
      operationId: getCurrentSession
      summary: Inspect the current browser session
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
              examples: {standard: {$ref: '#/components/examples/AuthResponse'}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      tags: [Authentication]
      operationId: logout
      summary: Revoke or clear the current browser session
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Session revoked or absent cookie cleared}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/auth/sessions:
    get:
      tags: [Authentication]
      operationId: listSessions
      summary: List the user's bounded session collection
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Sessions
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionCollection'}
              example:
                sessions:
                  - {id: 44444444-4444-4444-8444-444444444444, user_agent: Example Browser, created_at: '2026-08-27T09:00:00Z', last_seen_at: '2026-08-27T09:05:00Z', idle_expires_at: '2026-08-27T09:35:00Z', absolute_expires_at: '2026-08-27T21:00:00Z', current: true}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      tags: [Authentication]
      operationId: revokeAllSessions
      summary: Revoke all sessions for the user
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Sessions revoked}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/auth/sessions/{session_id}:
    delete:
      tags: [Authentication]
      operationId: revokeSession
      summary: Revoke one session belonging to the user
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/SessionID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {description: Session revoked}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/auth/password:
    put:
      tags: [Authentication]
      operationId: changePassword
      summary: Change the current user's password and rotate the session
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PasswordRequest'}
            example: {current_password: old-correct-password, new_password: new-correct-password}
      responses:
        '200':
          description: Password changed and session rotated
          headers:
            Set-Cookie: {$ref: '#/components/headers/SessionCookie'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
              examples: {standard: {$ref: '#/components/examples/AuthResponse'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      summary: List organizations visible to the user
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Organization page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrganizationPage'}
              example:
                items:
                  - {id: 11111111-1111-4111-8111-111111111111, name: Applied Systems Lab, lifecycle_state: active, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z', deleted_at: null}
                next_cursor: null
                previous_cursor: null
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Create an organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationWrite'}
            example: {name: Applied Systems Lab}
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Get an organization
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      summary: Rename an organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationWrite'}
            example: {name: Applied Systems Laboratory}
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      tags: [Organizations]
      operationId: deleteOrganization
      summary: Soft-delete an organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Organization deleted}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/restore:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    post:
      tags: [Organizations]
      operationId: restoreOrganization
      summary: Restore a soft-deleted organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '200':
          description: Organization restored
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/teams:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Teams]
      operationId: listTeams
      summary: List teams in an organization
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Team page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TeamPage'}
              example:
                items:
                  - {id: 22222222-2222-4222-8222-222222222222, name: Reproducibility Group, slug: reproducibility-group, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z'}
                next_cursor: null
                previous_cursor: null
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Teams]
      operationId: createTeam
      summary: Create a team
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TeamWrite'}
            example: {name: Reproducibility Group}
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Team'}
              examples: {standard: {$ref: '#/components/examples/Team'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/invitations:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      operationId: listOrganizationInvitations
      security: [{sessionCookie: []}]
      responses: {'200': {description: Pending invitations}}
    post:
      operationId: createOrganizationInvitation
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}}}
      responses: {'201': {description: Invitation created}}
  /api/v1/orgs/{org_id}/members:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      operationId: listOrganizationMembers
      security: [{sessionCookie: []}]
      responses: {'200': {description: Organization members}}
  /api/v1/orgs/{org_id}/invitations/{invitation_id}:
    parameters:
      - {$ref: '#/components/parameters/OrganizationID'}
      - {name: invitation_id, in: path, required: true, schema: {$ref: '#/components/schemas/PublicID'}}
    delete:
      operationId: revokeOrganizationInvitation
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses: {'204': {description: Invitation revoked}}
  /api/v1/organization-invitations/current:
    get:
      operationId: getCurrentOrganizationInvitation
      parameters: [{$ref: '#/components/parameters/InvitationToken'}]
      responses: {'200': {description: Invitation preview}}
    delete:
      operationId: declineCurrentOrganizationInvitation
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/InvitationToken'}
      responses: {'204': {description: Invitation declined}}
  /api/v1/organization-invitations/current/accept:
    post:
      operationId: acceptCurrentOrganizationInvitation
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/InvitationToken'}
      responses: {'200': {description: Invitation accepted}}
  /api/v1/orgs/{org_id}/teams/{team_id}/members:
    parameters:
      - {$ref: '#/components/parameters/OrganizationID'}
      - {$ref: '#/components/parameters/TeamID'}
    get:
      operationId: listTeamMembers
      security: [{sessionCookie: []}]
      responses: {'200': {description: Team members}}
    post:
      operationId: addTeamMember
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}}}
      responses: {'201': {description: Team member added}}
  /api/v1/orgs/{org_id}/teams/{team_id}/tasks:
    parameters:
      - {$ref: '#/components/parameters/OrganizationID'}
      - {$ref: '#/components/parameters/TeamID'}
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List tasks in a team
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Task page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskPage'}
              example:
                items:
                  - {id: 33333333-3333-4333-8333-333333333333, description: Reproduce benchmark results, status: todo, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z'}
                next_cursor: null
                previous_cursor: null
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a task
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskWrite'}
            example: {description: Reproduce benchmark results, status: todo}
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
              examples: {standard: {$ref: '#/components/examples/Task'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
  /api/v1/orgs/{org_id}/resources:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Resources]
      operationId: listResources
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Resources, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceCollection'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Resources]
      operationId: createResource
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceWrite'}}}}
      responses:
        '201': {description: Resource created, content: {application/json: {schema: {$ref: '#/components/schemas/Resource'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-synodus_session
      description: Production browser session cookie.
...
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 512}
...
    SessionID:
      name: session_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
...
  responses:
...
    Unauthenticated: {$ref: '#/components/responses/UnauthenticatedResponse'}
    Forbidden: {$ref: '#/components/responses/ForbiddenResponse'}
...
    RateLimited: {$ref: '#/components/responses/RateLimitedResponse'}
    InternalFailure: {$ref: '#/components/responses/InternalFailureResponse'}
    DependencyUnavailable: {$ref: '#/components/responses/DependencyUnavailableResponse'}
...
  schemas:
...
    AuthResponse:
      type: object
      additionalProperties: false
      required: [user, session, csrf_token]
      properties:
        user: {$ref: '#/components/schemas/User'}
        session: {$ref: '#/components/schemas/Session'}
        csrf_token: {type: string, minLength: 1, maxLength: 512}
    SessionCollection:
      type: object
      additionalProperties: false
      required: [sessions]
      properties:
        sessions:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/Session'}
...
  examples:
    AuthResponse:
      value:
        user: {id: 11111111-1111-4111-8111-111111111111, email: researcher@example.com, display_name: Researcher}
        session: {id: 44444444-4444-4444-8444-444444444444, created_at: '2026-08-27T09:00:00Z', last_seen_at: '2026-08-27T09:05:00Z', idle_expires_at: '2026-08-27T09:35:00Z', absolute_expires_at: '2026-08-27T21:00:00Z', current: true}
        csrf_token: example-csrf-token
...